requirements.txt
flask==2.0.1
werkzeug==2.0.2
web3==5.24.0
eth-account==0.5.7
python-dotenv==0.19.1
gunicorn==20.1.0
flask-cors==3.0.10
cryptography==36.0.0
requests==2.26.0 